+++ src/RazorLight/Compilation/TemplateCompilationException.cs
@@ -1,7 +1,5 @@
-﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using Microsoft.CodeAnalysis;
 
 namespace RazorLight.Compilation
 {
@@ -9,19 +7,10 @@ namespace RazorLight.Compilation
 	{
 		private readonly List<TemplateCompilationDiagnostic> compilationDiagnostics = new List<TemplateCompilationDiagnostic>();
 
-		public IReadOnlyList<string> CompilationErrors => compilationDiagnostics?.Select(x => x.FormattedMessage).ToList() ?? new List<string>();
+		public IReadOnlyList<string> CompilationErrors => compilationDiagnostics.Select(x => x.FormattedMessage).ToList();
 
 		public IReadOnlyList<TemplateCompilationDiagnostic> CompilationDiagnostics => compilationDiagnostics;
 
-		[Obsolete("Use constructor that takes enumerable of TemplateCompilationDiagnostic as input parameters")]
-		public TemplateCompilationException(string message, IEnumerable<string> errors) : base(message)
-		{
-			if (errors != null)
-			{
-				compilationDiagnostics.AddRange(errors.Select(x => new TemplateCompilationDiagnostic(x, x, null)));	
-			}
-		}
-		
 		public TemplateCompilationException(string message, IEnumerable<TemplateCompilationDiagnostic> diagnostics) : base(message)
 		{
 			if (diagnostics != null)