+++ src/RazorLight/IRazorLightEngineFactory.cs
@@ -1,9 +1,0 @@
-﻿using System.Reflection;
-
-namespace RazorLight
-{
-	public interface IRazorLightEngineFactory
-	{
-		IRazorLightEngine Create(Assembly? operatingAssembly = null, string? fileSystemProjectRoot = null);
-	}
-}